[constants] Rename files
