+++ src/lib/telemetry/TelemetryCollector.ts
@@ -3,7 +3,7 @@ import { ensureFileSync, writeJsonSync } from 'fs-extra'
 import { join } from 'path'
 import * as pkgJson from '../../../package.json'
 import logger from '../../logger'
-import { PathConstants } from '../constants/PathConstants'
+import { PathConstants } from '../constants/Paths'
 import { ErrorCreationArguments, ErrorReport, ErrorReportObj } from '../error/ErrorReport'
 import { Metric, MetricReport, MetricReportObj } from '../metrics/MetricReport'
 import { spawnUnblockingChildProcess } from '../utils/spawnUnblockingChildProcess'
